Adding end2end tests for bulk endpoint
Also improving metadata test, making each standalone.
